+++ doc/changelog.rst
@@ -38,6 +38,9 @@ New Features
 - Added :class:`~plydata.cat_tools.cat_infreq` which creates a categorical
   with categories in order of the number of times they appear in the sequence.
 
+- Added :class:`~plydata.cat_tools.cat_inseq` which creates a categorical
+  with categories in ascending numerical order.
+
 Enhancements
 ************